Add FC types to the index.
This would probably be better implemented by using different ES types,
but I think they are fundamentally represented the same way. In any
case, this way is a bit easier because we don't need to know the set
of possible FC types up front. If we used ES types, then we'd need to
probably expose them in the configuration.