Some minor bug fixes.
The first is related to creating mappings. Apparently, it seems that if
all documents for a type are deleting, then the mappings for that type
are also destroyed. This is important for tests, but probably not so
much for production. We fix this by decoupling index creation from
mapping creation.

The second is related to id sorting. For reasons beyond my
comprehension, Elasticsearch does not seem to do well when sorting by
the `_id` field. Instead, we opt for the `_uid` field, which various
enigmatic Internet posts suggest. (AIUI, `_id` doesn't actually exist,
only `_uid` does. `_id` is but an illusion!)