Pin elasticsearch to < 2.
Our tests currently rely on the ability to delete a type mapping in
Elasticsearch (and all documents associated with it), but this
functionality was removed from Elasticsearch 2.0 (and, correspondingly,
from the Python Elasticsearch 2.0 client).

This requires further investigation to fix. Deleting an entire index on
every test isn't really practical.